Scratch - CS50 - ElBeeL
My scratch project for CS50 is a game. It features my dog and my cat in outerspace. My cat is very protective of my dog and would likely jump in front of a meteor for her. They're best friends even in outer space.  

Use the arrow keys to move Ravioli the Kitten Bodyguard to defend Queen B in her Teepee from the meteors! (Block the meteors with Ravioli)
There are power-ups in the form of a toy-mouse and a cookie. Use them to your advantage and keep the Queen alive.

The longer you keep Queen B alive, the greater your reputation around the galaxy. Three hits and she's down for the count.